+++ inst/js_raws/webpack.config.js
@@ -1,5 +1,5 @@
-const webpack = require("webpack");
-const path = require('path');
+/* const webpack = require("webpack");
+// const path = require('path');
 const minimizer = require('terser-webpack-plugin');
 
 module.exports = {
@@ -17,4 +17,20 @@ module.exports = {
       include: /\.min\.js$/
     })]
   }
+};
+*/
+
+const path = require('path');
+
+module.exports = {
+  mode: 'production',
+  entry: path.resolve(__dirname, 'src/index.js'),
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, '../htmlwidgets/lib/dipterixThreeBrain-1.0.1'),
+    // libraryTarget: 'var',
+    // library: 'RAVEPipeline'
+    publicPath: "/"
+  },
+  devtool: 'source-map',
 };