+++ inst/js_raws/webpack.development.config.js
@@ -1,19 +1,13 @@
-const webpack = require("webpack");
 const path = require('path');
-const minimizer = require('terser-webpack-plugin');
 
 module.exports = {
-  mode: 'development',
   entry: path.resolve(__dirname, 'src/index.js'),
-  devtool: "source-map",
   output: {
+    filename: 'main.js',
     path: path.resolve(__dirname, '../htmlwidgets/lib/dipterixThreeBrain-1.0.1'),
-    filename: "main.js"
+    // libraryTarget: 'var',
+    // library: 'RAVEPipeline'
+    publicPath: "/"
   },
-  optimization: {
-    minimize: true,
-    minimizer: [new minimizer({
-      include: /\.min\.js$/
-    })]
-  }
+  devtool: 'source-map',
 };